Rename old Header to ContentHeader
