+++ packages/frontend/src/components/layouts/Container.tsx
@@ -1,0 +1,13 @@
+import clsx from 'clsx';
+import { FC } from 'react';
+
+type ContainerProps = {
+  className?: string;
+  children: React.ReactNode;
+};
+
+export const Container: FC<ContainerProps> = ({ className, children }) => {
+  return (
+    <div className={clsx('mx-auto max-w-screen-xl', className)}>{children}</div>
+  );
+};
