+++ packages/frontend/src/components/layouts/Layout.tsx
@@ -1,8 +1,8 @@
-import { Container } from './container/container';
+import { Container } from './Container';
 
 import { ToastContainer } from '$blocks/toast/toast.container';
+import { ContentHeader } from '$layouts/ContentHeader';
 import { Header } from '$layouts/Header';
-import { Header as SecondaryHeader } from '$layouts/header/header';
 
 type LayoutProps = {
   title: string;
@@ -19,7 +19,7 @@ export const Layout = ({ children, title }: LayoutProps): JSX.Element => {
             className="px-4 max-lg:mt-[64px] lg:mt-[82px] max-lg:pt-8 max-lg:pb-24 lg:px-8 lg:py-12"
             data-testid="layout-root"
           >
-            <SecondaryHeader title={title} />
+            <ContentHeader title={title} />
             <ToastContainer className="mb-8 -mt-2" />
             {children}
           </div>
